# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SiteCraft AI Website Generation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/generate:
    post:
      operationId: generateProject
      tags: [generation]
      summary: Start AI generation for a project
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "202":
          description: Generation job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIJob"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Generation already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}/chat-edit:
    post:
      operationId: chatEditProject
      tags: [generation]
      summary: Send a chat edit instruction to modify the project
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatEditRequest"
      responses:
        "202":
          description: Edit job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIJob"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/regenerate-section:
    post:
      operationId: regenerateSection
      tags: [generation]
      summary: Regenerate a specific section of the project
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateSectionRequest"
      responses:
        "202":
          description: Regeneration job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIJob"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /projects/{id}/deploy:
    post:
      operationId: deployProject
      tags: [deployments]
      summary: Deploy a project to Hostinger FTP
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeployRequest"
      responses:
        "202":
          description: Deployment started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{id}/deployments:
    get:
      operationId: listProjectDeployments
      tags: [deployments]
      summary: List deployments for a project
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentListResponse"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment by ID
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          $ref: "#/components/responses/NotFound"
  /deployments/{id}/rollback:
    post:
      operationId: rollbackDeployment
      tags: [deployments]
      summary: Rollback to this deployment
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "202":
          description: Rollback started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          $ref: "#/components/responses/NotFound"
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: List domains for current user
      responses:
        "200":
          description: List of domains
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainListResponse"

    post:
      operationId: createDomain
      tags: [domains]
      summary: Add a custom domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDomainRequest"
      responses:
        "201":
          description: Domain added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
  /domains/{id}:
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Remove a domain
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Domain removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    AIJob:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        type:
          type: string
          enum: [generate, chat-edit, regenerate-section]
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: number
          description: 0-100
        currentStep:
          type: string
          nullable: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/AIJobStep"
        error:
          type: string
          nullable: true
        result:
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - projectId
        - type
        - status
        - progress
        - steps
        - createdAt
        - updatedAt
    AIJobStep:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        order:
          type: integer
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        output:
          nullable: true
        error:
          type: string
          nullable: true
      required:
        - id
        - jobId
        - name
        - status
        - order
    GenerateRequest:
      type: object
      properties:
        businessDescription:
          type: string
        targetAudience:
          type: string
        primaryCta:
          type: string
        additionalInstructions:
          type: string
      required:
        - businessDescription
    ChatEditRequest:
      type: object
      properties:
        message:
          type: string
        contextJson:
          type: string
          nullable: true
          description: JSON-encoded context object
      required:
        - message
    RegenerateSectionRequest:
      type: object
      properties:
        sectionId:
          type: string
        instruction:
          type: string
      required:
        - sectionId
...
    Deployment:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        status:
          type: string
          enum: [pending, uploading, verifying, live, failed, rolled_back]
        environment:
          type: string
          enum: [production, staging]
        liveUrl:
          type: string
          nullable: true
        screenshotUrl:
          type: string
          nullable: true
        ftpHost:
          type: string
          nullable: true
        lighthouseScore:
          type: number
          nullable: true
        filesUploaded:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - projectId
        - status
        - environment
        - createdAt
    DeploymentListResponse:
      type: object
      properties:
        deployments:
          type: array
          items:
            $ref: "#/components/schemas/Deployment"
      required:
        - deployments
    DeployRequest:
      type: object
      properties:
        environment:
          type: string
          enum: [production, staging]
          default: production
        ftpHost:
          type: string
        ftpUsername:
          type: string
        ftpPassword:
          type: string
        ftpPath:
          type: string
      required:
        - ftpHost
        - ftpUsername
        - ftpPassword
    Domain:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        projectId:
          type: string
          nullable: true
        domain:
          type: string
        verified:
          type: boolean
        sslActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - domain
        - verified
        - sslActive
        - createdAt
    DomainListResponse:
      type: object
      properties:
        domains:
          type: array
          items:
            $ref: "#/components/schemas/Domain"
      required:
        - domains
    CreateDomainRequest:
      type: object
      properties:
        domain:
          type: string
        projectId:
          type: string
      required:
        - domain
...