# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SiteCraft AI Website Generation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Domain {
  id: string;
  userId: string;
  projectId?: string | null;
  domain: string;
  verified: boolean;
  sslActive: boolean;
  createdAt: Date;
}